casting into long

+++ src/com/simplilearn/calculator/ArithematicOperation.java
@@ -4,7 +4,7 @@ final class ArithematicOperation {
 	   float number1;
 	   float number2;
 	   char choice;
-	   private double result;
+	   private long result;
 	
 	  
 	public void Calculate(float number1, float number2,char choice) {
@@ -13,17 +13,17 @@ final class ArithematicOperation {
 		this.number2 = number2;
 		switch(choice) {
 		
-        case '+': result = number1 + number2;
+        case '+': result = (long) (number1 + number2);
         	System.out.println("Addition of "+number1+" and "+number2+"is :"+result);
            break;
        
-        case '-': result = number1 - number2;
+        case '-': result = (long) (number1 - number2);
         System.out.println("Substraction of "+number1+" and "+number2+"is :"+result);    break;
         
-        case '*': result = number1 * number2;
+        case '*': result = (long) (number1 * number2);
         System.out.println("Multiplication of "+number1+" and "+number2+"is :"+result);   break;
         
-        case '/': result = number1 / number2;
+        case '/': result = (long) (number1 / number2);
         System.out.println("Division of "+number1+" and "+number2+"is :"+result);   break;
         
         default: System.out.printf("Error! Please Enter correct operator");
